Feature/permissions

+++ app/Http/Controllers/TransactionController.php
@@ -9,6 +9,13 @@ use Illuminate\Http\Request;
 
 class TransactionController extends Controller
 {
+    private $request;
+
+    public function __construct(Request $request)
+    {
+        $this->request = $request;
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -16,8 +23,11 @@ class TransactionController extends Controller
      */
     public function index()
     {
-        $transactions = Transaction::paginate(10);
-        return $this->ok('Get all transaction success', $transactions);
+        $user = $this->request->user();
+        if ($user->can('SHOW_TRANSACTION')) {
+            $transactions = Transaction::paginate(10);
+            return $this->ok('Get all transaction success', $transactions);
+        }
     }
 
     /**
@@ -29,42 +39,47 @@ class TransactionController extends Controller
     public function store(Request $request)
     {
         $user = $request->user();
-        $input = $request->all();
+        if ($user->can('STORE_TRANSACTION')) {
 
-        $this->validate($request, [
-            'name' => 'required',
-            'amount' => 'required',
-            'type' => 'required',
-        ]);
+            $input = $request->all();
 
-        try {
-            DB::beginTransaction();
-            $transaction = new Transaction();
-            $transaction->name = $input['name'];
-            $transaction->amount = $input['amount'];
-            $wallet = Wallet::where('user_id', $user->id)->first();
-            if ($input['type'] == 'DEBIT') {
-                $wallet->ending_balance = $wallet->ending_balance + $transaction->amount;
-                $wallet->debit = $wallet->debit + $transaction->amount;
-            } else if ($input['type'] == 'CREDIT') {
-                $wallet->ending_balance = $wallet->ending_balance - $transaction->amount;
-                $wallet->credit = $wallet->credit + $transaction->amount;
-            } else {
-                return $this->badRequest('Transaction not available', $input);
+            $this->validate($request, [
+                'name' => 'required',
+                'amount' => 'required',
+                'type' => 'required',
+            ]);
+
+            try {
+                DB::beginTransaction();
+                $transaction = new Transaction();
+                $transaction->name = $input['name'];
+                $transaction->amount = $input['amount'];
+                $wallet = Wallet::where('user_id', $user->id)->first();
+                if ($input['type'] == 'DEBIT') {
+                    $wallet->ending_balance = $wallet->ending_balance + $transaction->amount;
+                    $wallet->debit = $wallet->debit + $transaction->amount;
+                } else if ($input['type'] == 'CREDIT') {
+                    $wallet->ending_balance = $wallet->ending_balance - $transaction->amount;
+                    $wallet->credit = $wallet->credit + $transaction->amount;
+                } else {
+                    return $this->badRequest('Transaction not available', $input);
+                }
+                $wallet->save();
+                $user->balance = $wallet->ending_balance;
+                $user->save();
+                $transaction->user_id = $user->id;
+                $transaction->wallet_id = $wallet->id;
+                $transaction->type = $input['type'];
+                $transaction->description = $transaction->name . ' with amount ' . $transaction->amount;
+                $transaction->save();
+                DB::commit();
+                return $this->created('Create transaction success', $transaction);
+            } catch (Exception $e) {
+                DB::rollBack();
+                return $this->internalServerError($e->getMessage());
             }
-            $wallet->save();
-            $user->balance = $wallet->ending_balance;
-            $user->save();
-            $transaction->user_id = $user->id;
-            $transaction->wallet_id = $wallet->id;
-            $transaction->type = $input['type'];
-            $transaction->description = $transaction->name . ' with amount ' . $transaction->amount;
-            $transaction->save();
-            DB::commit();
-            return $this->created('Create transaction success', $transaction);
-        } catch (Exception $e) {
-            DB::rollBack();
-            return $this->internalServerError('Error create transaction');
+        } else {
+            return $this->unauthorized();
         }
     }
 
@@ -74,13 +89,18 @@ class TransactionController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show(Request $request, $id)
     {
-        $transaction = Transaction::find($id);
-        if (!$transaction) {
-            return $this->badRequest('Transaction id ' . $id . ' not found');
+        $user = $request->user();
+        if ($user->can('SHOW_TRANSACTION')) {
+            $transaction = Transaction::where('id', $id)->where('user_id', $user->id)->first();
+            if (!$transaction) {
+                return $this->badRequest('Transaction id ' . $id . ' not found');
+            }
+            return $this->ok('Get transaction by id success', $transaction);
+        } else {
+            return $this->unauthorized();
         }
-        return $this->ok('Get transaction by id success', $transaction);
     }
 
     /**
@@ -93,43 +113,47 @@ class TransactionController extends Controller
     public function update(Request $request, $id)
     {
         $user = $request->user();
-        $input = $request->all();
+        if ($user->can('STORE_TRANSACTION')) {
+            $input = $request->all();
 
-        $this->validate($request, [
-            'type' => 'required',
-        ]);
+            $this->validate($request, [
+                'type' => 'required',
+            ]);
 
-        try {
-            DB::beginTransaction();
-            $transaction = Transaction::find($id);
-            if (!$transaction) {
-                return $this->badRequest('Transaction id ' . $id . ' not found');
+            try {
+                DB::beginTransaction();
+                $transaction = Transaction::where('id', $id)->where('user_id', $user->id)->first();
+                if (!$transaction) {
+                    return $this->badRequest('Transaction id ' . $id . ' not found');
+                }
+                $transaction->name = $input['name'];
+                $transaction->amount = $input['amount'];
+                $wallet = Wallet::where('user_id', $user->id)->first();
+                if ($input['type'] == 'DEBIT') {
+                    $wallet->ending_balance = $wallet->ending_balance + $transaction->amount;
+                    $wallet->debit = $wallet->debit + $transaction->amount;
+                } else if ($input['type'] == 'CREDIT') {
+                    $wallet->ending_balance = $wallet->ending_balance - $transaction->amount;
+                    $wallet->credit = $wallet->credit + $transaction->amount;
+                } else {
+                    return $this->badRequest('Transaction not available', $input);
+                }
+                $wallet->save();
+                $user->balance = $wallet->ending_balance;
+                $user->save();
+                $transaction->user_id = $user->id;
+                $transaction->wallet_id = $wallet->id;
+                $transaction->type = $input['type'];
+                $transaction->description = $transaction->name . ' with amount ' . $transaction->amount;
+                $transaction->save();
+                DB::commit();
+                return $this->created('Update transaction success', $transaction);
+            } catch (Exception $e) {
+                DB::rollBack();
+                return $this->internalServerError('Error update transaction');
             }
-            $transaction->name = $input['name'];
-            $transaction->amount = $input['amount'];
-            $wallet = Wallet::where('user_id', $user->id)->first();
-            if ($input['type'] == 'DEBIT') {
-                $wallet->ending_balance = $wallet->ending_balance + $transaction->amount;
-                $wallet->debit = $wallet->debit + $transaction->amount;
-            } else if ($input['type'] == 'CREDIT') {
-                $wallet->ending_balance = $wallet->ending_balance - $transaction->amount;
-                $wallet->credit = $wallet->credit + $transaction->amount;
-            } else {
-                return $this->badRequest('Transaction not available', $input);
-            }
-            $wallet->save();
-            $user->balance = $wallet->ending_balance;
-            $user->save();
-            $transaction->user_id = $user->id;
-            $transaction->wallet_id = $wallet->id;
-            $transaction->type = $input['type'];
-            $transaction->description = $transaction->name . ' with amount ' . $transaction->amount;
-            $transaction->save();
-            DB::commit();
-            return $this->created('Update transaction success', $transaction);
-        } catch (Exception $e) {
-            DB::rollBack();
-            return $this->internalServerError('Error update transaction');
+        } else {
+            return $this->unauthorized();
         }
     }
 
@@ -141,11 +165,16 @@ class TransactionController extends Controller
      */
     public function destroy($id)
     {
-        $transaction = Transaction::find($id);
-        if (!$transaction) {
-            return $this->badRequest('Transaction id ' . $id . ' not found');
+        $user = $this->request->user();
+        if ($user->can('DESTROY_TRANSACTION')) {
+            $transaction = Transaction::where('id', $id)->where('user_id', $user->id)->first();
+            if (!$transaction) {
+                return $this->badRequest('Transaction id ' . $id . ' not found');
+            }
+            $transaction->delete();
+            return $this->ok('Delete transaction success');
+        } else {
+            return $this->unauthorized();
         }
-        $transaction->delete();
-        return $this->ok('Delete transaction success');
     }
 }
